wip: mod manager tab [macata #94]
